+++ api/src/main/java/org/grnet/cat/api/endpoints/AssessmentsEndpoint.java
@@ -175,9 +175,9 @@ public class AssessmentsEndpoint {
     public Response getAssessment(@Parameter(
             description = "The ID of the assessment to retrieve.",
             required = true,
-            example = "1",
-            schema = @Schema(type = SchemaType.NUMBER)) @PathParam("id")
-                                  @Valid @NotFoundEntity(repository = AssessmentRepository.class, message = "There is no Assessment with the following id:") Long id) {
+            example = "c242e43f-9869-4fb0-b881-631bc5746ec0",
+            schema = @Schema(type = SchemaType.STRING)) @PathParam("id")
+                                  @Valid @NotFoundEntity(repository = AssessmentRepository.class, message = "There is no Assessment with the following id:") String id) {
 
         var validations = assessmentService.getDtoAssessment(utility.getUserUniqueIdentifier(), id);
 
@@ -233,8 +233,8 @@ public class AssessmentsEndpoint {
     public Response updateAssessment(@Parameter(
             description = "The ID of the assessment to update.",
             required = true,
-            example = "1",
-            schema = @Schema(type = SchemaType.NUMBER))
+            example = "c242e43f-9869-4fb0-b881-631bc5746ec0",
+            schema = @Schema(type = SchemaType.STRING))
                                      @PathParam("id")
                                      @Valid @NotFoundEntity(repository = AssessmentRepository.class, message = "There is no assessment with the following id:") String id,
                                      @Valid @NotNull(message = "The request body is empty.") JsonAssessmentRequest updateJsonAssessmentRequest) {