+++ entity/src/main/java/org/grnet/cat/entities/Assessment.java
@@ -1,7 +1,10 @@
 package org.grnet.cat.entities;
 
+
 import jakarta.persistence.*;
 import jakarta.validation.constraints.NotNull;
+import org.hibernate.annotations.UuidGenerator;
+
 import java.sql.Timestamp;
 
 /**
@@ -15,8 +18,8 @@ public class Assessment {
      * As id
      */
     @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    @UuidGenerator
+    private String id;
 
     @Column(name = "assessment_doc", columnDefinition = "json")
     private String assessmentDoc;
@@ -44,7 +47,7 @@ public class Assessment {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(String id) {
         this.id = id;
     }